Created update 1.5.0
Changed the low-current deadband threshold to activate if charge current is less than the threshold instead of less than/equal to (2023-01-01).
Updated copyright string in About screen to read "(C) 2021-2023".
Changed the default cell count to 2S for improved user experience; most PPS adapters go to 11V so a compatibility fail out of the box kinda sucks... (2023-01-06).
Changed how aggressive GC is enabled/disabled; set aggressiveGcThreshold to 0 instead of isAggressiveGcEnabled to false (not that you should do this anyway...) (2023-01-08).
Fixed issue where configuration menu libraries remain resident in memory even when no longer needed (2023-01-21).
Added LM135/LM235/LM335 support as external temperature sensors (2023-01-21).
Split off charge control function into a separate file which unloads upon termination to conserve memory (2023-01-27).
Renamed "DC4S-CompileMenu" to "DC4S-CompileLibs" to reflect that non-menu libraries are also compiled here (2023-01-27).
Changed how USB-C CC attachment errors are handled; user can retry the detection instead of needing to restart the charge setup procedure (2023-01-28).
