+++ source/lua/user/DC4S/lib/DC4S-cfgCableRes.lua
@@ -2,7 +2,8 @@
 https://ripitapart.com December 15, 2022.
 
 Version history:
-1.4.0: Split off monolithic menu library functions into individual files (2022-12-15).]]
+1.4.0: Split off monolithic menu library functions into individual files (2022-12-15).
+1.5.0: Fixed issue where configuration menu libraries remain resident in memory even when no longer needed (2023-01-21).]]
 
 function cfgCableRes()
   screen.clear()
@@ -40,5 +41,7 @@ function cfgCableRes()
   -- discard temporary variables
   resSel = nil
   tmpRes = nil
+  cfgCableRes = nil
+  package.loaded["lua/user/DC4S/lib/DC4S-cfgCableRes"] = nil
   collectgarbage("collect") -- clean up memory
 end
