+++ source/lua/user/DC4S/lib/DC4S-cfgPChgCRate.lua
@@ -2,7 +2,8 @@
 https://ripitapart.com December 15, 2022.
 
 Version history:
-1.4.0: Split off monolithic menu library functions into individual files (2022-12-15).]]
+1.4.0: Split off monolithic menu library functions into individual files (2022-12-15).
+1.5.0: Fixed issue where configuration menu libraries remain resident in memory even when no longer needed (2023-01-21).]]
 
 function cfgPChgCRate()
   screen.clear()
@@ -21,5 +22,7 @@ function cfgPChgCRate()
   -- discard temporary variables
   tRateSel = nil
   tmpPCRate = nil
+  cfgPChgCRate = nil
+  package.loaded["lua/user/DC4S/lib/DC4S-cfgPChgCRate"] = nil
   collectgarbage("collect") -- clean up memory
 end
