+++ source/lua/user/DC4S/lib/DC4S-cfgPreChg.lua
@@ -2,7 +2,8 @@
 https://ripitapart.com December 15, 2022.
 
 Version history:
-1.4.0: Split off monolithic menu library functions into individual files (2022-12-15).]]
+1.4.0: Split off monolithic menu library functions into individual files (2022-12-15).
+1.5.0: Fixed issue where configuration menu libraries remain resident in memory even when no longer needed (2023-01-21).]]
 
 function cfgPreChg()
   local preChgSel = 0
@@ -27,5 +28,7 @@ function cfgPreChg()
   end
   -- discard temporary variables
   preChgSel = nil
+  cfgPreChg = nil
+  package.loaded["lua/user/DC4S/lib/DC4S-cfgPreChg"] = nil
   collectgarbage("collect") -- clean up memory
 end
