+++ source/lua/user/DC4S/lib/DC4S-cfgSounds.lua
@@ -2,7 +2,8 @@
 https://ripitapart.com December 15, 2022.
 
 Version history:
-1.4.0: Split off monolithic menu library functions into individual files (2022-12-15).]]
+1.4.0: Split off monolithic menu library functions into individual files (2022-12-15).
+1.5.0: Fixed issue where configuration menu libraries remain resident in memory even when no longer needed (2023-01-21).]]
 
 function cfgSounds()
   local cfgSoundSel = 0
@@ -39,5 +40,7 @@ function cfgSounds()
   -- discard temporary variables
   cfgSoundSel = nil
   soundMenuValue = nil
+  cfgSounds = nil
+  package.loaded["lua/user/DC4S/lib/DC4S-cfgSounds"] = nil
   collectgarbage("collect") -- clean up memory
 end
